added test that runs the notebook
