+++ tests/test_as_example.py
@@ -3,27 +3,6 @@ import pytest
 import numpy as np
 
 
-@pytest.fixture
-def check_reg(num_regression, data_regression):
-
-    def _check(jf):
-        num_regression.check({
-            'psd': jf.psd,
-            'logpsdK': jf.dct.logpsdK,
-            'logpsdK_THEORY_std': jf.dct.logpsdK_THEORY_std,
-            'logtau': jf.dct.logtau,
-            'logtau_THEORY_std': jf.dct.logtau_THEORY_std,
-            'KAPPA_SCALE': np.array([float(jf.KAPPA_SCALE)]),
-            'Nyquist_f_THz': np.array([float(jf.Nyquist_f_THz)]),
-            'kappa_Kmin': np.array([float(jf.kappa_Kmin)]),
-            'kappa_Kmin_std': np.array([float(jf.kappa_Kmin_std)]),
-            'aic_min': np.array([float(jf.dct.aic_min)])
-        })
-        data_regression.check({'aic_Kmin': int(jf.dct.aic_Kmin),})
-
-    return _check
-
-
 def test_example_NaCl(data_NaCl, check_reg):
     import thermocepstrum as tc
