+++ tests/test_notebooks/final_result1.txt
@@ -1,0 +1,9 @@
+-----------------------------------------------------
+  CEPSTRAL ANALYSIS
+-----------------------------------------------------
+  AIC_Kmin  = 3  (P* = 4, corr_factor = 1.000000)
+  L_0*   =          15.158757 +/-   0.056227
+  S_0*   =     6824108.702608 +/- 383697.095268
+-----------------------------------------------------
+  kappa* =           0.498310 +/-   0.028018  W/m/K
+-----------------------------------------------------
