+++ tests/test_notebooks/final_result2.txt
@@ -1,0 +1,9 @@
+-----------------------------------------------------
+  CEPSTRAL ANALYSIS
+-----------------------------------------------------
+  AIC_Kmin  = 37  (P* = 38, corr_factor = 1.000000)
+  L_0*   =          13.395456 +/-   0.103277
+  S_0*   =      950203.538450 +/- 98134.481881
+-----------------------------------------------------
+  kappa* =           2.484534 +/-   0.256596  W/m/K
+-----------------------------------------------------
